move notification viewer list, fix click detection, layout tweaks
